Fix test_compiler
